Move indexer stuff to separate files.
