+++ ledger/eval_test.go
@@ -1677,152 +1677,6 @@ func TestModifiedAppLocalStates(t *testing.T) {
 	}
 }
 
-type indexerLedgerForEvalImpl struct {
-	l           *Ledger
-	latestRound basics.Round
-}
-
-func (il indexerLedgerForEvalImpl) LatestBlockHdr() (bookkeeping.BlockHeader, error) {
-	return il.l.BlockHdr(il.latestRound)
-}
-
-// The value of the returned map is nil iff the account was not found.
-func (il indexerLedgerForEvalImpl) LookupWithoutRewards(addresses map[basics.Address]struct{}) (map[basics.Address]*basics.AccountData, error) {
-	res := make(map[basics.Address]*basics.AccountData)
-
-	for address := range addresses {
-		accountData, _, err := il.l.LookupWithoutRewards(il.latestRound, address)
-		if err != nil {
-			return nil, err
-		}
-
-		if accountData.IsZero() {
-			res[address] = nil
-		} else {
-			accountDataCopy := new(basics.AccountData)
-			*accountDataCopy = accountData
-			res[address] = accountDataCopy
-		}
-	}
-
-	return res, nil
-}
-
-func (il indexerLedgerForEvalImpl) GetAssetCreator(map[basics.AssetIndex]struct{}) (map[basics.AssetIndex]FoundAddress, error) {
-	// This function is unused.
-	return nil, errors.New("GetAssetCreator() not implemented")
-}
-
-func (il indexerLedgerForEvalImpl) GetAppCreator(map[basics.AppIndex]struct{}) (map[basics.AppIndex]FoundAddress, error) {
-	// This function is unused.
-	return nil, errors.New("GetAppCreator() not implemented")
-}
-
-func (il indexerLedgerForEvalImpl) Totals() (ledgercore.AccountTotals, error) {
-	return il.l.Totals(il.latestRound)
-}
-
-// Test that overriding the consensus parameters effects the generated apply data.
-func TestCustomProtocolParams(t *testing.T) {
-	partitiontest.PartitionTest(t)
-
-	genesisBalances, addrs, _ := newTestGenesis()
-
-	var genHash crypto.Digest
-	crypto.RandBytes(genHash[:])
-	block, err := bookkeeping.MakeGenesisBlock(protocol.ConsensusV24,
-		genesisBalances, "test", genHash)
-
-	dbName := fmt.Sprintf("%s", t.Name())
-	cfg := config.GetDefaultLocal()
-	cfg.Archival = true
-	l, err := OpenLedger(logging.Base(), dbName, true, InitState{
-		Block:       block,
-		Accounts:    genesisBalances.Balances,
-		GenesisHash: genHash,
-	}, cfg)
-	require.NoError(t, err)
-	defer l.Close()
-
-	const assetid basics.AssetIndex = 1
-	proto := config.Consensus[protocol.ConsensusV24]
-
-	block = bookkeeping.MakeBlock(block.BlockHeader)
-
-	createTxn := txntest.Txn{
-		Type:        "acfg",
-		Sender:      addrs[0],
-		GenesisHash: block.GenesisHash(),
-		AssetParams: basics.AssetParams{
-			Total:    200,
-			Decimals: 0,
-			Manager:  addrs[0],
-			Reserve:  addrs[0],
-			Freeze:   addrs[0],
-			Clawback: addrs[0],
-		},
-	}
-	createTxn.FillDefaults(proto)
-	createStib, err := block.BlockHeader.EncodeSignedTxn(
-		createTxn.SignedTxn(), transactions.ApplyData{})
-	require.NoError(t, err)
-
-	optInTxn := txntest.Txn{
-		Type:          "axfer",
-		Sender:        addrs[1],
-		GenesisHash:   block.GenesisHash(),
-		XferAsset:     assetid,
-		AssetAmount:   0,
-		AssetReceiver: addrs[1],
-	}
-	optInTxn.FillDefaults(proto)
-	optInStib, err := block.BlockHeader.EncodeSignedTxn(
-		optInTxn.SignedTxn(), transactions.ApplyData{})
-	require.NoError(t, err)
-
-	fundTxn := txntest.Txn{
-		Type:          "axfer",
-		Sender:        addrs[0],
-		GenesisHash:   block.GenesisHash(),
-		XferAsset:     assetid,
-		AssetAmount:   100,
-		AssetReceiver: addrs[1],
-	}
-	fundTxn.FillDefaults(proto)
-	fundStib, err := block.BlockHeader.EncodeSignedTxn(
-		fundTxn.SignedTxn(), transactions.ApplyData{})
-	require.NoError(t, err)
-
-	optOutTxn := txntest.Txn{
-		Type:          "axfer",
-		Sender:        addrs[1],
-		GenesisHash:   block.GenesisHash(),
-		XferAsset:     assetid,
-		AssetAmount:   30,
-		AssetReceiver: addrs[0],
-		AssetCloseTo:  addrs[0],
-	}
-	optOutTxn.FillDefaults(proto)
-	optOutStib, err := block.BlockHeader.EncodeSignedTxn(
-		optOutTxn.SignedTxn(), transactions.ApplyData{})
-	require.NoError(t, err)
-
-	block.Payset = []transactions.SignedTxnInBlock{
-		createStib, optInStib, fundStib, optOutStib,
-	}
-
-	il := indexerLedgerForEvalImpl{
-		l:           l,
-		latestRound: 0,
-	}
-	proto.EnableAssetCloseAmount = true
-	_, modifiedTxns, err := EvalForIndexer(il, &block, proto)
-	require.NoError(t, err)
-
-	require.Equal(t, 4, len(modifiedTxns))
-	assert.Equal(t, uint64(70), modifiedTxns[3].AssetClosingAmount)
-}
-
 // TestAppInsMinBalance checks that accounts with MaxAppsOptedIn are accepted by block evaluator
 // and do not cause any MaximumMinimumBalance problems
 func TestAppInsMinBalance(t *testing.T) {
